銘柄表をコンポーネントに切り出した
